+++ CHANGELOG.md
@@ -15,6 +15,13 @@ and this project partially adheres to [Semantic Versioning](https://semver.org/s
 
 ### Deprecated
 
+## [0.24] - 2022-09-04
+## Fixed
+- Inline Wiki Close Button
+
+## Added
+- Inline Wiki Entry for Band of Toughness
+
 ## [0.23] - 2022-09-03
 ## Added
 - Experimental Inline Wiki (currently only Aglet of the Wind)